+++ lib/view/profile_screen/profile_screen.dart
@@ -5,7 +5,6 @@ import 'package:cheat_sheet/res/typo.dart';
 import 'package:cheat_sheet/view/profile_screen/profile_sub_page/my_sheet.dart';
 import 'package:cheat_sheet/view/profile_screen/profile_sub_page/buy_sheet.dart';
 
-
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 
@@ -19,7 +18,7 @@ class ProfileScreen extends StatefulWidget {
 }
 
 class _ProfileScreenState extends State<ProfileScreen>
-  with AutomaticKeepAliveClientMixin, TickerProviderStateMixin {
+    with AutomaticKeepAliveClientMixin, TickerProviderStateMixin {
   late TabController tabController;
   final _auth = FirebaseAuth.instance;
   final _firestore = FirebaseFirestore.instance;
@@ -42,12 +41,19 @@ class _ProfileScreenState extends State<ProfileScreen>
     double screenHeight = MediaQuery.of(context).size.height;
 
     return StreamBuilder<DocumentSnapshot<Map<String, dynamic>>>(
-      stream: _firestore.collection("users").doc(_auth.currentUser?.uid).snapshots(),
-      builder: (BuildContext context, AsyncSnapshot<DocumentSnapshot> snapshot) {
-        if (!snapshot.hasData) {
-          return const Center(child: CircularProgressIndicator(),);
-        } else {
-            Map<String, dynamic> data = snapshot.data!.data() as Map<String, dynamic>;
+        stream: _firestore
+            .collection("users")
+            .doc(_auth.currentUser?.uid)
+            .snapshots(),
+        builder:
+            (BuildContext context, AsyncSnapshot<DocumentSnapshot> snapshot) {
+          if (!snapshot.hasData) {
+            return const Center(
+              child: CircularProgressIndicator(),
+            );
+          } else {
+            Map<String, dynamic> data =
+                snapshot.data!.data() as Map<String, dynamic>;
             return Scaffold(
               body: SafeArea(
                 child: DefaultTabController(
@@ -78,7 +84,8 @@ class _ProfileScreenState extends State<ProfileScreen>
                                           child: Container(
                                             decoration: BoxDecoration(
                                               image: DecorationImage(
-                                                image: AssetImage(data['profileImage']),
+                                                image: AssetImage(
+                                                    data['profileImage']),
                                                 // fit: BoxFit.fill,
                                               ),
                                               border: Border.all(
@@ -101,8 +108,8 @@ class _ProfileScreenState extends State<ProfileScreen>
                                                   MainAxisAlignment.center,
                                               children: <Widget>[
                                                 Medium24px(
-                                                  // text: _auth.currentUser!.displayName.toString()),
-                                                  text: data['username']),
+                                                    // text: _auth.currentUser!.displayName.toString()),
+                                                    text: data['username']),
                                                 IconButton(
                                                   icon: Icon(Icons.edit),
                                                   onPressed: () {},
@@ -121,14 +128,18 @@ class _ProfileScreenState extends State<ProfileScreen>
                                               children: [
                                                 Column(
                                                   children: [
-                                                    Medium20px(text: data['follower']),
+                                                    Medium20px(
+                                                        text: data['follower']
+                                                            .toString()),
                                                     const Regular14px(
                                                         text: "ผู้ติดตาม"),
                                                   ],
                                                 ),
                                                 Column(
                                                   children: [
-                                                    Medium20px(text: data['following']),
+                                                    Medium20px(
+                                                        text: data['following']
+                                                            .toString()),
                                                     const Regular14px(
                                                         text: "กำลังติตดาม"),
                                                   ],
@@ -142,7 +153,6 @@ class _ProfileScreenState extends State<ProfileScreen>
                                   );
                                 },
                               ),
-
                             ),
                             TabBar(
                               controller: tabController,
@@ -150,7 +160,6 @@ class _ProfileScreenState extends State<ProfileScreen>
                                 fontSize: 16,
                                 fontFamily: 'BaiJamjuree',
                                 fontWeight: FontWeight.w500,
-
                               ),
                               labelColor: AppColors.primary700,
                               indicatorWeight: 1,
