+++ lib/view/review_sub_page/_5star.dart
@@ -8,7 +8,7 @@ class FiveStar extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return ListView.builder(
-      itemCount: 5,
+      itemCount: 2,
       shrinkWrap: true,
       itemBuilder: (context, index) {
         return Review(userRating: 5);
