+++ lib/view/review_sub_page/all.dart
@@ -2,13 +2,13 @@ import 'package:flutter/material.dart';
 
 import '../../res/components/review.dart';
 
-class AllReview extends StatelessWidget {
-  const AllReview({super.key});
+class All extends StatelessWidget {
+  const All({super.key});
 
   @override
   Widget build(BuildContext context) {
     return ListView.builder(
-      itemCount: 6,
+      itemCount: 2,
       shrinkWrap: true,
       itemBuilder: (context, index) {
         return Review(userRating: 5);
